Libratized the project - pulled our library functionality out of the in-line server implementation and into a separate library folder. Invoked our own library functions to implement the library inline in our example web page.
