+++ server/server.js
@@ -2,21 +2,25 @@
 const express = require('express');
 const app = express();
 const server = require('http').Server(app); // eslint-disable-line
-const io = require('socket.io')(server);
+// const io = require('socket.io')(server);
 const path = require('path');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 const useragent = require('express-useragent');
-
-const activeConnectRequests = {};
-const desktopController = require('./desktopController.js');
-const mobileController = require('./mobileController.js');
+const echo = require('./../library/server/mainServer.js')(server); // ********
+console.log('init method: ', echo.init);
+console.log('return of init method: ', echo.init());
+// const activeConnectRequests = {};
+// const desktopController = require('./desktopController.js');
+// const mobileController = require('./mobileController.js');
 
 app.use(express.static(path.join(`${__dirname}/../client`)));
-app.use(useragent.express());
-app.use(bodyParser.urlencoded( { extended: true } ));
+app.use(useragent.express()); // TODO tie this into our library somehow?
+app.use(bodyParser.urlencoded({ extended: true }));
 app.use(cookieParser());
 app.set('view engine', 'ejs');
+// app.use((req, res, next) => { console.log('echo in middleware: ', echo); next(); });
+app.use(echo.init()); // ***************************
 
 /* ------------------
  * --    Routes    --
@@ -24,8 +28,6 @@ app.set('view engine', 'ejs');
 
 // App will serve up different pages for client & desktop
 app.get('/',
-  desktopController.handleRequest(activeConnectRequests),
-  mobileController.handleRequest(activeConnectRequests),
   (req, res) => {
     if (req.useragent && req.useragent.isDesktop) {
       res.sendFile(path.join(`${__dirname}/../client/browser.html`));
@@ -38,7 +40,6 @@ app.get('/',
   }
 );
 app.post('/',
-  mobileController.handlePost(activeConnectRequests),
   (req, res) => {
     if (req.useragent && req.useragent.isMobile) {
       // TODO Validate nonce match, if it doesn't, serve rootmobile
@@ -60,30 +61,30 @@ app.get('*', (req, res) => {
  * --   Sockets    --
  * ------------------ */
 
-io.on('connection', socket => {
-  console.log('A socket has a connection');
-  socket.on('createRoom', room => {
-    console.log(`Joined ${room}`);
-    // decrypt token?
-    socket.join(room);
-  });
-  socket.on('tap', room => {
-    console.log('Tap from mobile!');
-    io.sockets.in(room).emit('tap');
-  });
-  socket.on('disconnect', () => {
-    console.log('A user has disconnected');
-    io.emit('user disconnected');
-  });
-  socket.on('acceleration', (room, accObject) => {
-    console.log(accObject);
-    io.sockets.in(room).emit('acceleration', accObject);
-  });
-  socket.on('gyroscope', (room, gyroObject) => {
-    console.log(gyroObject);
-    io.sockets.in(room).emit('gyroscope', gyroObject);
-  });
-});
+// io.on('connection', socket => {
+//   console.log('A socket has a connection');
+//   socket.on('createRoom', room => {
+//     console.log(`Joined ${room}`);
+//     // decrypt token?
+//     socket.join(room);
+//   });
+//   socket.on('tap', room => {
+//     console.log('Tap from mobile!');
+//     io.sockets.in(room).emit('tap');
+//   });
+//   socket.on('disconnect', () => {
+//     console.log('A user has disconnected');
+//     io.emit('user disconnected');
+//   });
+//   socket.on('acceleration', (room, accObject) => {
+//     console.log(accObject);
+//     io.sockets.in(room).emit('acceleration', accObject);
+//   });
+//   socket.on('gyroscope', (room, gyroObject) => {
+//     console.log(gyroObject);
+//     io.sockets.in(room).emit('gyroscope', gyroObject);
+//   });
+// });
 
 /* ------------------
  * --    Server    --
